Icicle: Source: clean up; split queries into modules
